[frontend]junu_GNB영역 및 detailpage
