+++ frontend/assets/Buttons/EditButton.js
@@ -1,0 +1,8 @@
+export default `<svg xmlns="http://www.w3.org/2000/svg" width="26" height="26" viewBox="0 0 26 26">
+    <g transform="translate(-25 -19)">
+        <path data-name="사각형 21" transform="translate(25 19)" style="fill:none" d="M0 0h26v26H0z"/>
+        <circle data-name="타원 18" cx="1.5" cy="1.5" r="1.5" transform="translate(44 23)" style="fill:#fff"/>
+        <circle data-name="타원 19" cx="1.5" cy="1.5" r="1.5" transform="translate(44 31)" style="fill:#fff"/>
+        <circle data-name="타원 20" cx="1.5" cy="1.5" r="1.5" transform="translate(44 39)" style="fill:#fff"/>
+    </g>
+</svg>`;
