+++ frontend/assets/Buttons/backButton.js
@@ -1,30 +1,5 @@
-export default `<svg width="56" height="55" viewBox="0 0 56 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-<g clip-path="url(#clip0_2_1515)">
-<g filter="url(#filter0_d_2_1515)">
-<ellipse cx="27.1559" cy="27.8438" rx="26.8439" ry="27" fill="url(#paint0_linear_2_1515)"/>
-</g>
-<path d="M24.8265 29.1163C24.6697 28.9702 24.5453 28.7964 24.4603 28.6049C24.3754 28.4134 24.3317 28.208 24.3317 28.0006C24.3317 27.7932 24.3754 27.5878 24.4603 27.3963C24.5453 27.2048 24.6697 27.031 24.8265 26.885L32.5052 19.6882C32.662 19.5422 32.7864 19.3684 32.8713 19.1769C32.9563 18.9854 33 18.78 33 18.5726C33 18.3651 32.9563 18.1598 32.8713 17.9683C32.7864 17.7768 32.662 17.603 32.5052 17.4569C32.1917 17.1643 31.7677 17 31.3258 17C30.8838 17 30.4598 17.1643 30.1464 17.4569L22.4677 24.6694C21.5279 25.5533 21 26.7514 21 28.0006C21 29.2498 21.5279 30.448 22.4677 31.3318L30.1464 38.5443C30.458 38.8346 30.8785 38.9982 31.3174 39C31.5376 39.0012 31.7558 38.9615 31.9596 38.8833C32.1635 38.8051 32.3488 38.6899 32.5052 38.5443C32.662 38.3982 32.7864 38.2244 32.8713 38.0329C32.9563 37.8415 33 37.6361 33 37.4286C33 37.2212 32.9563 37.0158 32.8713 36.8243C32.7864 36.6329 32.662 36.4591 32.5052 36.313L24.8265 29.1163Z" fill="#FFEBFF"/>
-</g>
-<defs>
-<filter id="filter0_d_2_1515" x="-3.68799" y="0.84375" width="61.6879" height="62" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-<feFlood flood-opacity="0" result="BackgroundImageFix"/>
-<feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
-<feOffset dy="4"/>
-<feGaussianBlur stdDeviation="2"/>
-<feComposite in2="hardAlpha" operator="out"/>
-<feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"/>
-<feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_2_1515"/>
-<feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_2_1515" result="shape"/>
-</filter>
-<linearGradient id="paint0_linear_2_1515" x1="53.9999" y1="0.843749" x2="40.5468" y2="63.4847" gradientUnits="userSpaceOnUse">
-<stop stop-color="#A66E9F"/>
-<stop offset="0.494901" stop-color="#922A92"/>
-<stop offset="0.495001" stop-color="#922A92"/>
-<stop offset="0.817612" stop-color="#7D2D9D"/>
-<stop offset="1" stop-color="#5A3084"/>
-</linearGradient>
-<clipPath id="clip0_2_1515">
-<rect width="56" height="55" fill="white"/>
-</clipPath>
-</defs>
+export default `<svg xmlns="http://www.w3.org/2000/svg" width="26" height="26" viewBox="0 0 26 26">
+    <path data-name="사각형 21" style="fill:none" d="M0 0h26v26H0z"/>
+    <path data-name="패스 50" d="m40.9 40.9-8 8 8 8" transform="translate(-32 -35.898)" style="stroke-linejoin:round;stroke:#fff;stroke-linecap:round;stroke-width:1.5px;fill:none"/>
+    <path data-name="패스 51" d="M31.219 48h20" transform="translate(-30 -35)" style="stroke:#fff;stroke-linecap:round;stroke-width:1.5px;fill:none"/>
 </svg>`;
