+++ frontend/src/features/map/screen/component/dropPreview.js
@@ -21,7 +21,7 @@ import { Favourite } from "../../../../components/favorites/favourite.component"
 import { useState } from "react";
 import { theme } from "../../../../infrastructure/theme";
 import { FlatList } from "react-native-gesture-handler";
-import { DetailScreen } from "../DetailScreen";
+import { DetailScreen } from "../../../detail/DetailScreen";
 
 export const DropPreview = ({
   pressedLocation = {},
