+++ frontend/src/services/local.js
@@ -1,5 +1,5 @@
 //###.###.#.## 부분에 localhost 주소를 입력해 주세요.
-export default LOCAL_HOST = "192.168.0.2";
+export default LOCAL_HOST = "172.20.10.2";
 
 // 1. 터미널에서 ipconfig 입력
 // 2. 하단의 무선 LAN 어댑터 Wi-Fi 중 IPv4 주소 입력.
